Allow sharing of arbitrary files
We now allow the user to provide a mime-type they want to share. With
this, they can share arbitrary files, like audio files, documents or
whatever they can find an app for that wants this file type to be shared
with.

For this we introduce a new method share_file() to the Godot api and
aditionally augment the SharedData class with an extra mime_type field.
To keep the current API plugin API stable, data files are still shared
via set_image_path(), but can be any files. If no mime-type is
provided, we will default to text/plain if no file is provided and
image/* if a file is provided.

Fixes #5