Refactor Godot interface to use share_file()
The code of all share_*() methods that end up sharing a file now use
share_file() plus a mime-type for sharing. This also moves the mime-type
handling outside of the Java part of the code, leaving only the fallback
of text/plain in there, if no mime-type is provided by the user of the
plugin. share_text() is left the way it is, as it doesn't share a file
and doing extra handling for empty string in share_file() felt a bit
hacky (null is not allowed as String in Godot's typing system).

This commit includes a small behavioral change: We are now creating an
image in share_viewport() and share_texture(), even if the plugin has
not been initialized. This shouldn't be a problem, as having an
initialized plugin should be our default path.