Rename image_path to file_path in plugin
As we are now not only capable of sharing images, but any file type
whatsoever, we rename all plugin methods named after "image" to "file",
including the dictionary keys used to transfer data from the Godot side
to the Java side.

This will not be visible to anyone using the share_*() methods, but will
break code that is using the share plugin directly.