Avoid startup error message on non-Android systems
When running a Godot application with the share plugin enabled on a
non-Android system we always get an error message:

ERROR: SharePlugin singleton not found!

As we don't have a share plugin on non-Android platform, we don't need
to try to instantiate the share plugin. We do this by only calling
_update_plugin() when we're on Android.