Allow sharing of arbitrary files (#8)
* Allow sharing of arbitrary files

We now allow the user to provide a mime-type they want to share. With
this, they can share arbitrary files, like audio files, documents or
whatever they can find an app for that wants this file type to be shared
with.

For this we introduce a new method share_file() to the Godot api and
aditionally augment the SharedData class with an extra mime_type field.
To keep the current API plugin API stable, data files are still shared
via set_image_path(), but can be any files. If no mime-type is
provided, we will default to text/plain if no file is provided and
image/* if a file is provided.

Fixes #5

* Refactor Godot interface to use share_file()

The code of all share_*() methods that end up sharing a file now use
share_file() plus a mime-type for sharing. This also moves the mime-type
handling outside of the Java part of the code, leaving only the fallback
of text/plain in there, if no mime-type is provided by the user of the
plugin. share_text() is left the way it is, as it doesn't share a file
and doing extra handling for empty string in share_file() felt a bit
hacky (null is not allowed as String in Godot's typing system).

This commit includes a small behavioral change: We are now creating an
image in share_viewport() and share_texture(), even if the plugin has
not been initialized. This shouldn't be a problem, as having an
initialized plugin should be our default path.

* Rename image_path to file_path in plugin

As we are now not only capable of sharing images, but any file type
whatsoever, we rename all plugin methods named after "image" to "file",
including the dictionary keys used to transfer data from the Godot side
to the Java side.

This will not be visible to anyone using the share_*() methods, but will
break code that is using the share plugin directly.